adapted theme to better match rest of program
