+++ AdviceGiverListPlaceholder.HTML
@@ -1,0 +1,67 @@
+
+<html>
+    <head>
+        <meta charset="utf-8">
+        <script src="AdviceGiverList-SCRIPT.js"></script>
+        
+        <style>
+    
+    body{
+        font-family:Arial, Helvetica, sans-serif;
+        max-width:600px; 
+        height: 900px; 
+        border: 1px solid black;
+    }
+    
+    header{
+            text-align:center;
+            background-color:#009688;
+            color: white;
+            height: 50px;
+        }
+        #wrapper{
+            margin: 50px 50px;
+            height: 50px;
+            width: 500px;
+            background-color: #009688;
+        }
+    
+        #intro{
+            background-color:peachpuff;
+            text-align: center;
+            font-weight:bold;
+            border: solid 5px #009688;
+        }
+    
+        #profile_data{
+            margin: 30px;
+            background-color: bisque;
+            border: solid 2px black;
+        }
+    
+        
+    
+        #button {
+        font-size:          1.2em;
+        background-color:   #BBBBBB;
+        margin:             12px;
+        width:              10em;
+        margin-left: 125px;
+        background-color: #009688
+            } </style>
+    </head>
+    
+    <body>
+        
+      <header>
+          Placeholder html page
+      </header>
+      <form action = "http://localhost:8080">
+      <button id="button" >"Ask Advice" /</button>
+    </form>
+   
+        
+    </body>
+    
+    </html>
+
